Move post related containers to specified folded
