+++ frontend/src/components/Post/Page.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Post from "../../containers/Post";
+import Post from "../../containers/Post/Post";
 import CommentList from "../Comment/List";
 
 const Page = ({ postId, commentIds }) => (
