+++ frontend/src/containers/Post/Post.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import Post from "../components/Post";
+import Post from "../../components/Post";
 
 const mapStateToProps = ({ posts }, { id }) => ({
   post: posts[id] || null
